+++ corpus/hook-ns-clash/a/src/a/core.clj
@@ -1,0 +1,6 @@
+(ns a.core)
+
+(defn foo [x] x)
+
+(defn use-it []
+  (foo 1))