+++ corpus/hook-ns-clash/b/src/b/core.clj
@@ -1,0 +1,6 @@
+(ns b.core)
+
+(defn bar [x] x)
+
+(defn use-it []
+  (bar 1))